Adds support for bibliography tags in layouts (#202)
Specifically addresses the use case of expected a bibiliography for
every page that uses a layout. This patch modifies the cited_keys
function so that the backing array store is attached to the page's data.
This allows the layout to see the cited references when it renders the
bibliography (with the `--cited` option). Previously this did not work
because the cited_keys array was attached to the Jekyll context, which
is different between the layout and the page.