Remove use_raw_bibtex_entry option
Jekyll 4 does not render Liquid markup from included BibTeX files any
longer. This obviates the need for adding `{%raw%}` around an entry's
`bibtex` field. Indeed, adding it causes `{%raw%}` and `{%endraw%}` to
appear in the output.

This commit adds a warning if the `use_raw_bibtex_entry` key appears in
the jekyll-scholar options, regardless of the value of the entry.

Jekyll 4 requires at least Ruby 2.4 so increase the version numbers for
Travis.