Update test to work correctly with removing duplicates.
This test needs to be updated.  The reason is that two entries a, and b,
are duplicates as implemented by bibtex-ruby since the only difference
is the key.  The key does not distinguish entries as per bibtex-ruby.

Change the year to 2018 for entry b.

The second test also must be updated for a similar reason. The two
entries are the same.

The grouping test is the same.  The entries are regarded as duplicates
by bibtex-ruby, and should change.