Use uniq! to remove duplicates (#276)
* Supporting textit is straightforward.  Simply use emph.

* Update the feature test.

* Remove print statement.

* Enable textit by default

* Update tests

* Add lowercase test.

* Passing lowercase test.

* Add textregistered latex to HTML filter.

* Clean up debug statements.

* Use uniq! to remove duplicates by default from bibliography object.

* Update test to work correctly with removing duplicates.

This test needs to be updated.  The reason is that two entries a, and b,
are duplicates as implemented by bibtex-ruby since the only difference
is the key.  The key does not distinguish entries as per bibtex-ruby.

Change the year to 2018 for entry b.

The second test also must be updated for a similar reason. The two
entries are the same.

The grouping test is the same.  The entries are regarded as duplicates
by bibtex-ruby, and should change.

* By default remove_duplicates is false. (issue #276)

Updated some tests as well.

Configuring what bibtex-ruby considers a duplicate is not provided yet.

* By default remove_duplicates is false. (issue #276)

Updated some tests as well.

Configuring what bibtex-ruby considers a duplicate is not provided yet.

* Add options to remove duplicates using '-r'.