Use page front matter to override configuration
Closes #95

Turns out, this is incredibly simple (2 lines changed! :smile:).
When Jekyll passes a context to the `render` method, it has a register
with the hash `:page` which bundles all the front matter defined for the
given page. To make life easier, the front matter configuration matches
the schema for the `scholar` configuration in `_config.yml` exactly.

Now, simply merge the configuration hashes!